# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Budget API",
        "description": "API endpoints for managing budgets",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/suppliers": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get all suppliers",
                "operationId": "e9e2528153818477e1fb8cc7c607aabd",
                "responses": {
                    "200": {
                        "description": "List of suppliers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Supplier"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create a new supplier",
                "operationId": "5a8fe28248cbaf0665bb42be5fe2a2df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SupplierRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/suppliers/{id}": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get supplier by ID",
                "operationId": "2f5be2f63f47faff6ae5c8ef276a3b2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Update an existing supplier",
                "operationId": "2492a53ecf96a8603b11ebf46400a88f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SupplierRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Delete a supplier",
                "operationId": "87da2a2246c3842089cc9a6672778241",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supplier deleted successfully"
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Supplier": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "person": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "last_name": {
                                "type": "string"
                            },
                            "address": {
                                "type": "string"
                            },
                            "phone_number": {
                                "type": "string"
                            },
                            "mail": {
                                "type": "string"
                            },
                            "dni": {
                                "type": "string"
                            },
                            "cuit": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SupplierRequest": {
                "required": [
                    "notes",
                    "person_id",
                    "person"
                ],
                "properties": {
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "person_id": {
                        "type": "integer"
                    },
                    "person": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "last_name": {
                                "type": "string"
                            },
                            "address": {
                                "type": "string"
                            },
                            "phone_number": {
                                "type": "string"
                            },
                            "mail": {
                                "type": "string"
                            },
                            "dni": {
                                "type": "string"
                            },
                            "cuit": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}